Ensure that 'iocage exec' treats 'cmd' as opaque
When the string provided to 'cmd' for the 'exec' mode
contains options which match iocage's options, iocage
absorbs them instead of passing them along. This patch
adds '--' in the correct position in the 'iocage exec'
command string to ensure that iocage treats
everything afterwards as the command to be executed.

Signed-off-by: Kevin P. Fleming <kevin@km6g.us>